+++ pygbif/__init__.py
@@ -36,12 +36,12 @@ Usage::
 		occurrences.search(geometry='POLYGON((30.1 10.1, 10 20, 20 40, 40 40, 30.1 10.1))', limit=20)
 '''
 
-__version__ = '0.1.6.0'
+__version__ = '0.2.0'
 __title__ = 'pygbif'
 __author__ = 'Scott Chamberlain'
 __license__ = 'MIT'
 
 from .occurrences import search, get, count, download
-from .species import name_parser, name_suggest, name_backbone, name_lookup
+from .species import name_parser, name_suggest, name_backbone, name_lookup, name_usage
 from .registry import datasets, nodes, networks, organizations, installations
 from .gbifissues import occ_issues_lookup