+++ test/test-registry-datasets.py
@@ -1,4 +1,4 @@
-"""Tests for registry module - datasets methods"""
+"""Tests for registry module - datasets method"""
 import os
 from pygbif import registry
 
@@ -25,15 +25,3 @@ def test_datasets_type():
     assert 100 == len(res['results'])
     assert 'OCCURRENCE' == list(set(vv))[0]
 
-def test_dataset_metrics():
-    "registry.dataset_metrics - basic test"
-    res = registry.dataset_metrics(uuid='3f8a1297-3259-4700-91fc-acc4170b27ce')
-    assert dict == res.__class__
-    assert 19 == len(res)
-
-def test_dataset_metrics_multiple_uuids():
-    "registry.dataset_metrics - multiple uuids"
-    uuids = ['3f8a1297-3259-4700-91fc-acc4170b27ce', '66dd0960-2d7d-46ee-a491-87b9adcfe7b1']
-    res = registry.dataset_metrics(uuids)
-    assert list == res.__class__
-    assert 2 == len(res)